fix: change ForwardBodyConfig.MaxSize from uint16 to int32 (#5099)
* fix: change ForwardBodyConfig.MaxSize from uint16 to int32

uint16 is not supported by go-to-protobuf; int32 aligns with other
Gateway API integer fields and fixes protobuf conversion for
ForwardBodyConfig.

* Update apis/v1/httproute_types.go


* Regenerate CRDs after adding Minimum=0 to ForwardBodyConfig.MaxSize

---------

+++ apis/v1/httproute_types.go
@@ -1791,7 +1791,8 @@ type ForwardBodyConfig struct {
 	//
 	// If 0, the body will not be sent to the authorization server.
 	// +optional
-	MaxSize uint16 `json:"maxSize,omitempty"`
+	// +kubebuilder:validation:Minimum=0
+	MaxSize int32 `json:"maxSize,omitempty"`
 }
 
 // HTTPBackendRef defines how a HTTPRoute forwards a HTTP request.
